+++ CHANGELOG.md
@@ -30,6 +30,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - (_reducers_): `getStateByAction` and `getStateByNamespace` have been moved to the `@redux-tools/namespaces` package.
 - (_injectors-react_): It is no longer necessary to pass `isGlobal: true` when not using the namespacing mechanism.
 
+### Fixed
+
+- (_namespaces-react_): Static namespace and feature in `namespacedConnect` is no longer passed down to inner components.
+
 ### Removed
 
 - The unpkg bundle is no longer created for non-preset packages.